requirements.txt
# Minimal Google GenAI + LangChain setup with Pinecone
# Core LangChain - pinned to known working versions
langchain==0.2.16
langchain-core==0.2.39
langchain-community==0.2.16
langchain-text-splitters==0.2.4

# Google GenAI integration - pinned to stable version
langchain-google-genai==1.0.10

# Google AI SDK
google-generativeai==0.7.2

# Pinecone integration - compatible versions for LangChain 0.2.x
pinecone-client==3.2.2
langchain-pinecone==0.1.1
pinecone-text==0.7.1

# Pydantic - compatible version (critical for avoiding schema errors)
pydantic==2.8.2
pydantic-core==2.20.1

# Core dependencies
typing-extensions==4.12.2
annotated-types==0.6.0

# HTTP clients
httpx==0.27.0
requests==2.32.3

# Async support
anyio==4.4.0
sniffio==1.3.1

# JSON handling
orjson==3.10.7

# Document processing (minimal)
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3

# Vector store (if needed) - Older compatible version with pinecone-text
chromadb==0.4.15
chroma-hnswlib==0.7.3

# Streamlit
streamlit==1.37.1

# FastAPI (minimal version)
fastapi==0.111.1
uvicorn==0.30.5

# Utilities
python-dotenv==1.0.1
numpy>=1.21.5,<=1.25.2
pandas==2.2.2

# Authentication
google-auth==2.34.0

# System utilities
psutil==6.0.0
packaging==24.1
click==8.1.7

# File handling
filelock==3.15.4

# Date/time
python-dateutil==2.9.0.post0

# Encoding
charset-normalizer==3.3.2
idna==3.8
certifi==2024.8.30
urllib3==2.2.2

# Caching
cachetools==5.4.0

# Protobuf (for Google APIs)
protobuf==4.25.4
googleapis-common-protos==1.63.2
grpcio==1.65.4

# Additional core deps
six==1.16.0
PyYAML==6.0.2
Jinja2==3.1.4
MarkupSafe==2.1.5

# Machine Learning
scikit-learn==1.3.2
scipy>=1.3.2,<1.12.0

# SQLite3 binary for Linux
pysqlite3-binary==0.5.3; sys_platform == 'linux'